Updated to add padding
Issues sometimes get created with a formatting error - the environment info gets turned into a heading if the next line contains the three dashes. The solution is to add a blank line between the environment info and the three dashes. Adding an extra blank line here might help prevent this.