Update issue templates with checklist
I've turned the checklist at the bottom of the issue templates into an
actual checklist - in GitHub and Codetree, you will be able to see if
all items have been checked off or not.

I've updated the language a little bit as well to be clearer/more
precise.

I have added a note about including a testing plan in the comments of
the issue, but this could maybe be stronger and/or handled differently -
open to suggestions.

I've divided documentation and release notes into two separate
requirements as sometimes one or the other are not really required.

The goal of this is to properly meet our definition of done for issues
before they are closed, because once an issue is closed no one ever
looks at it again.