Adds security and title notes to template
The security note points reporters to the SECURITY.md file in the repo
for more info on how to safely report security issues.

The title note asks reporters to title the issue as a problem statement.
We could perhaps add more info to this note, but I wanted to keep it
brief.